# src/main/java/ru/skypro/avito/dto/LoginReq.java
package ru.skypro.avito.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import lombok.Data;

import javax.validation.constraints.Email;
import javax.validation.constraints.NotBlank;
import javax.validation.constraints.Size;

import static ru.skypro.avito.constant.Regexp.EMAIL_REGEXP;

@Data
public class LoginReq {
    @Email(regexp = EMAIL_REGEXP)
    @Schema(example = "user@user.ru")
    private String username;
    @NotBlank
    @Size(min = 8)
    private String password;
}